+++ src/components/Block/Block.styled.ts
@@ -16,34 +16,32 @@ export const OuterWrapper = styled.div<{ padding?: boolean }>`
     padding-left: ${(props) => (props.padding ? '1.5rem' : '0')};
     padding-right: ${(props) => (props.padding ? '1.5rem' : '0')};
     box-sizing: border-box;
-    border: 8px dashed rgb(96, 108, 243);
 `;
 
 export const ContentWrapper = styled.div`
     display: flex;
     flex-direction: column;
     align-items: center;
-    border: 2px dashed red;
 `;
 
 export const TitleText = styled.div`
     font-size: 1.8rem;
     text-align: center;
-    font-weight: 600;
+    font-weight: 640;
     color: ${({ theme }: { theme: { theme: keyof typeof themeSetting } }) => themeSetting[theme.theme].titleTextColor};
 `;
 
 export const SubtitleText = styled.div`
-    width: 80%;
+    width: 70%;
     text-align: center;
     margin-top: 0.2rem;
     font-size: 1.4rem;
-    font-weight: 300;
+    font-weight: 360;
     color: ${({ theme }: { theme: { theme: keyof typeof themeSetting } }) =>
         themeSetting[theme.theme].subtitleTextColor};
 `;
 
 export const Content = styled.div`
-    margin-top: 1.4rem;
+    margin-top: 3rem;
     font-size: var(--default-font-size);
 `;