+++ src/pages/Home/Home.tsx
@@ -1,9 +1,9 @@
-// import { Block } from 'src/components/Block/Block';
+import { Block } from 'src/components/Block/Block';
 import { OuterWrapper } from './Home.styled';
 // import { MoreProducts } from 'src/components/MoreProducts/MoreProducts';
 // import { ContactUs } from 'src/components/ContactUs/ContactUs';
 // import { OurPages } from 'src/components/OurPages/OurPages';
-// import { FeaturedItems } from 'src/components/FeaturedItems/FeaturedItems';
+import { FeaturedItems } from 'src/components/FeaturedItems/FeaturedItems';
 import { HomePageHeader } from 'src/components/HomePageHeader/HomePageHeader';
 
 /**
@@ -14,12 +14,10 @@ export const Home = () => {
     return (
         <OuterWrapper>
             <HomePageHeader />
-
-            {/* <Block title="Featured Items" subtitle="New arrivals & top products from our sponsors" padding>
-                block
+            <Block title="Featured Items" subtitle="New arrivals & top products from our sponsors" padding>
+                <FeaturedItems />
             </Block>
-            <
-            <Block title="Featured products" padding>
+            {/*<Block title="Featured products" padding>
                 <MoreProducts />
             </Block>
             <Block title="Contact us" padding>